Use pooch to download/cache fixtures, for now just new ones
Refs #117
