+++ pyproject.toml
@@ -73,6 +73,7 @@ type = [
 ]
 dev = [
     "aiohttp-compress>=0.2.1",
+    "pooch>=1.9.0",
     "pytest>=9.0.3",
     "pytest-aiohttp>=1.1.0",
     "pytest-asyncio>=1.1.0",
